+++ TDEngine2/include/ecs/CUIEventsSystem.h
@@ -87,6 +87,9 @@ namespace TDEngine2
 
 			IInputContext* mpInputContext; 
 			IDesktopInputContext* mpDesktopInputContext;
+
+			std::string mInputBuffer;
+			std::string mPrevInputBuffer;
 	};