+++ TDEngine2/plugins/UnixInputContext/source/CUnixInputContext.cpp
@@ -33,12 +33,10 @@ namespace TDEngine2
 		mpDisplayHandler = pUnixWindowSystem->GetInternalData().mpDisplayHandler;
 		mWindowHandler   = pUnixWindowSystem->GetInternalData().mWindowHandler;
 
-#if TDE2_EDITORS_ENABLED /// This subscription is needed only with editor mode's enabled
 		if (auto pEventManager = pWindowSystem->GetEventManager())
 		{
 			pEventManager->Subscribe(TOnCharInputEvent::GetTypeId(), this);
 		}
-#endif
 
 		mIsInitialized = true;
 
@@ -113,15 +111,11 @@ namespace TDEngine2
 
 	E_RESULT_CODE CUnixInputContext::OnEvent(const TBaseEvent* pEvent)
 	{
-#if TDE2_EDITORS_ENABLED
-		if (mOnCharInputCallback)
+		if (auto pInputEvent = dynamic_cast<const TOnCharInputEvent*>(pEvent))
 		{
-			if (auto pInputEvent = dynamic_cast<const TOnCharInputEvent*>(pEvent))
-			{
-				mOnCharInputCallback(static_cast<TUtf8CodePoint>(pInputEvent->mCharCode));
-			}
+			mOnCharInputCallback(static_cast<TUtf8CodePoint>(pInputEvent->mCharCode));
 		}
-#endif
+
 		return RC_OK;
 	}
 
@@ -166,15 +160,11 @@ namespace TDEngine2
         return TPtr<IGamepad>(nullptr);
 	}
 
-#if TDE2_EDITORS_ENABLED
-
 	void CUnixInputContext::SetOnCharInputCallback(const TOnCharActionCallback& onEventAction)
 	{
-		mOnCharInputCallback = onEventAction;
+		mOnCharInputCallback.Subscribe(onEventAction);
 	}
 	
-#endif
-
 	U8 CUnixInputContext::_toKeycode(E_KEYCODES keyCode)
 	{
 		switch (keyCode)