+++ TDEngine2/source/platform/win32/CWin32WindowSystem.cpp
@@ -55,9 +55,7 @@ namespace TDEngine2
 
 		};
 
-#if TDE2_EDITORS_ENABLED
 		TOnCharInputEvent onCharInputEvent;
-#endif
 
 		switch (uMsg)
 		{
@@ -92,12 +90,10 @@ namespace TDEngine2
 				pEventManager->Notify(&onMovedEvent);
 				break;
 
-#if TDE2_EDITORS_ENABLED
 			case WM_CHAR:
 				onCharInputEvent.mCharCode = static_cast<U32>(wParam);
 				pEventManager->Notify(&onCharInputEvent);
 				break;
-#endif
 
 			default:
 				return DefWindowProc(hWnd, uMsg, wParam, lParam);